+++ src/Client/ClientCore/Network/Pipeline/Inbound/ClientAllowedEventsInboundPipelineStage.cs
@@ -36,7 +36,8 @@ public class ClientAllowedEventsInboundPipelineStage : IInboundPipelineStage
         EventId.Core_Ping_Ping,
         EventId.Core_WorldManagement_Subscribe,
         EventId.Core_WorldManagement_Unsubscribe,
-        EventId.Client_EntitySynchronization_Update
+        EventId.Client_EntitySynchronization_Update,
+        EventId.Core_Movement_Move
     };
 
     public ILogger Logger { private get; set; } = NullLogger.Instance;