+++ src/Common/NetworkCore/Network/Pipeline/Outbound/DeliveryMethodOutboundPipelineStage.cs
@@ -35,7 +35,9 @@ public class DeliveryMethodOutboundPipelineStage : IOutboundPipelineStage
         { EventId.Core_Ping_Pong, DeliveryMethod.Unreliable },
         { EventId.Core_WorldManagement_Subscribe, DeliveryMethod.ReliableUnordered },
         { EventId.Core_WorldManagement_Unsubscribe, DeliveryMethod.ReliableUnordered },
-        { EventId.Client_EntitySynchronization_Update, DeliveryMethod.ReliableUnordered }
+        { EventId.Client_EntitySynchronization_Update, DeliveryMethod.ReliableUnordered },
+        { EventId.Core_Movement_Move, DeliveryMethod.ReliableUnordered },
+        { EventId.Core_Movement_RequestMove, DeliveryMethod.ReliableUnordered }
     };
 
     public ILogger Logger { private get; set; } = NullLogger.Instance;